ITRDB extraction updates
Work with JSON format using jsonlite makes life SO MUCH EASIER!

This makes for a cleaner code and a TON easier to find useful metadata.
 This also means we don’t need to make a new data frame with the
metadata since we can convert the JSON info to a data frame.

Next steps will be adding support for non-square bounding box &
converting to a full function